ScrappingBot/requirements.txt
pytest==7.4.3
selenium==4.14.0
pytest-html==4.1.1
pandas==2.1.3
Unidecode==1.3.7
pymysql==1.1.0
mysql-connector-python==8.2.0
